Advances in camera

+++ Base/Motor2D/Render.cpp
@@ -148,11 +148,12 @@ pair<int,int> Render::ScreenToWorld(int x, int y) const
 bool Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, SDL_RendererFlip flip, float speed, double angle, int pivot_x, int pivot_y) const
 {
 	bool ret = true;
-	uint scale = App->win->GetScale();
+	// Scale must be always positive (changed from uint to int)
+	int scale = App->win->GetScale();
 
 	SDL_Rect rect;
-	rect.x = (int)(camera.x * speed) + x * scale * zoom;
-	rect.y = (int)(camera.y * speed) + y * scale * zoom;
+	rect.x = (int)( ((camera.x * speed) + (int)((x * scale)*zoom)) );
+	rect.y = (int)( ((camera.y * speed) + (int)((y * scale)*zoom)) );
 
 	if(section != NULL)
 	{
